+++ src/tabs/createDC_culLife.js
